requirements.txt
aniso8601==7.0.0
autopep8==1.4.4
certifi==2019.9.11
chardet==3.0.4
coverage==4.5.4
Django==2.2.5
django-filter==2.2.0
django-graphql-jwt==0.2.2
factory-boy==2.12.0
Faker==2.0.2
graphene==2.1.8
graphene-django==2.5.0
graphql-core==2.2.1
graphql-relay==2.0.0
idna==2.8
promise==2.2.1
psycopg2==2.8.3
pycodestyle==2.5.0
PyJWT==1.7.1
python-coveralls==2.9.3
python-dateutil==2.8.0
pytz==2019.2
PyYAML==5.1.2
requests==2.22.0
Rx==1.6.1
singledispatch==3.4.0.3
six==1.12.0
sqlparse==0.3.0
text-unidecode==1.3
urllib3==1.25.3
